Changed to dotnet 7 aswell as moved offer to shared project
