+++ SharedObjects/SharedObjects.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>SharedObjects</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Offer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Recipe.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)User.cs" />
   </ItemGroup>
